Separated functions into their own separate files
